// web/src/types/mypage.ts
export type RoutePreview = {
  id: string;
  title: string;
  meta: string;
  savedAt: string;
  thumbnail: string;
};

export type PostPreview = {
  id: string;
  title: string;
  excerpt: string;
  date: string;
  like: number;
  comments: number;
};

export type CertPreview = {
  id: string;
  title: string;
  place: string;
  datetime: string;
  thumbnail: string;
};